feat(today): TODAY 잠금 섹션 경험과 내비게이션을 개선 (#77)
