+++ src/pages/home/ui/home-status-feedback.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useEffect } from "react";
+import { useEffect, useRef } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 
 import { toast } from "sonner";
@@ -10,15 +10,25 @@ export function HomeStatusFeedback() {
   const router = useRouter();
   const searchParams = useSearchParams();
   const safePathname = pathname ?? "/home";
+  const hasHandledLockedToastRef = useRef(false);
 
   useEffect(() => {
     const todayState = searchParams?.get("today");
 
     if (todayState !== "locked") {
+      hasHandledLockedToastRef.current = false;
       return;
     }
 
-    toast.info("오늘 글을 작성한 뒤 TODAY를 볼 수 있어요.");
+    if (hasHandledLockedToastRef.current) {
+      return;
+    }
+
+    hasHandledLockedToastRef.current = true;
+
+    toast.info("오늘 글을 작성한 뒤 TODAY를 볼 수 있어요.", {
+      id: "today-locked-feedback",
+    });
 
     const nextSearchParams = new URLSearchParams(searchParams?.toString());
     nextSearchParams.delete("today");
