--- src/pages/landing/model/index.ts
@@ -1,1 +1,0 @@
-export { mockPosts } from "./mock-posts";
